Introduce `DeviceMetadata`
